requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
c2pa-python==0.6.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
idna==3.10
logging==0.4.9.6
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.10.1
pywin32==308
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
